refactor(components): tidy file placement and naming
- delete models/Tooltip.tsx (dead re-export of ui/tooltip)
- rename models/Models.tsx → models.tsx (kebab-case)
- rename keys/delete-confirmation.tsx → key-delete-confirmation.tsx
- move model-utils, model-categories, modality-ui from keys/ to models/ (they are model-domain helpers, not key UI)
- also: button.tsx → ui/, faq.tsx + news-banner.tsx → news-faq/
