+++ enter.pollinations.ai/src/client/components/internal/design-showcase.tsx
@@ -3,10 +3,10 @@ import { cn } from "@/util.ts";
 import { Chart } from "../activity/chart.tsx";
 import { MultiSelect } from "../activity/multi-select.tsx";
 import type { DataPoint } from "../activity/types.ts";
-import { Button } from "../button.tsx";
 import { ModalityButton } from "../keys/modality-button.tsx";
-import { MODALITY_COLORS, type Modality } from "../keys/modality-ui.ts";
 import { type ThemeName, themes } from "../layout/dashboard-theme.ts";
+import { MODALITY_COLORS, type Modality } from "../models/modality-ui.ts";
+import { Button } from "../ui/button.tsx";
 import { Chip } from "../ui/chip.tsx";
 import { IconButton } from "../ui/icon-button.tsx";
 import { InfoTip } from "../ui/info-tip.tsx";
