+++ enter.pollinations.ai/src/client/components/keys/api-key-list.tsx
@@ -8,8 +8,8 @@ import { IconButton } from "../ui/icon-button.tsx";
 import { Surface } from "../ui/surface.tsx";
 import { Tooltip } from "../ui/tooltip.tsx";
 import { ApiKeyDialog } from "./api-key-dialog.tsx";
-import { DeleteConfirmation } from "./delete-confirmation.tsx";
 import { EditApiKeyDialog } from "./edit-api-key-dialog.tsx";
+import { DeleteConfirmation } from "./key-delete-confirmation.tsx";
 import { KeyDisplay } from "./key-display.tsx";
 import { LimitsBadge, shortLocale } from "./limits-badge.tsx";
 import { ModelsBadge } from "./models-badge.tsx";
