+++ enter.pollinations.ai/src/client/components/keys/publishable-key-settings.tsx
@@ -1,6 +1,6 @@
 import { Field } from "@ark-ui/react/field";
 import type { FC } from "react";
-import { Button } from "../button.tsx";
+import { Button } from "../ui/button.tsx";
 import { Switch } from "../ui/switch.tsx";
 
 type PublishableKeySettingsProps = {
