--- enter.pollinations.ai/src/client/components/models/Tooltip.tsx
@@ -1,2 +1,0 @@
-// Re-export from shared ui location
-export { Tooltip } from "../ui/tooltip.tsx";
