+++ enter.pollinations.ai/src/client/components/models/index.ts
@@ -1,2 +1,2 @@
-export { Models } from "./Models.tsx";
+export { Models } from "./models.tsx";
 export type { Modalities, ModelPrice } from "./types.ts";
