+++ enter.pollinations.ai/src/client/components/news-faq/faq.tsx
@@ -2,8 +2,8 @@ import type { FC } from "react";
 import { useEffect, useState } from "react";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
-import faqMarkdown from "../../../POLLEN_FAQ.md?raw";
-import { PollenExamples } from "./models/pollen-examples.tsx";
+import faqMarkdown from "../../../../POLLEN_FAQ.md?raw";
+import { PollenExamples } from "../models/pollen-examples.tsx";
 
 export const FAQ_GITHUB_URL =
     "https://github.com/pollinations/pollinations/blob/master/enter.pollinations.ai/POLLEN_FAQ.md";
