+++ enter.pollinations.ai/src/client/components/news-faq/news-banner.tsx
