+++ enter.pollinations.ai/src/client/components/pollen/tier-explanation.tsx
@@ -5,8 +5,8 @@ import {
 } from "@shared/tier-config.ts";
 import type { FC, ReactNode } from "react";
 import { cn } from "@/util.ts";
-import { Tooltip } from "../models/Tooltip.tsx";
 import { Surface } from "../ui/surface.tsx";
+import { Tooltip } from "../ui/tooltip.tsx";
 
 const SeedTooltipContent = () => (
     <div className="w-72">
