+++ enter.pollinations.ai/src/client/components/ui/link-button.tsx
@@ -1,7 +1,7 @@
 import type { FC, ReactNode } from "react";
 import { cn } from "@/util.ts";
-import { Button } from "../button.tsx";
 import type { ThemeName } from "../layout/dashboard-theme.ts";
+import { Button } from "./button.tsx";
 
 type LinkButtonProps = {
     theme: ThemeName;
